+++ CHANGELOG.rst
@@ -4,9 +4,9 @@ Changelog
 
 Notable changes to this project will be documented in this file.
 
-********
-Upcoming
-********
+*******************
+v2.5.3 (2023-11-19)
+*******************
 
 Added
 =====